refactor: 패키지 구조 대규모 변경
- DDD 기반 패키지 구조로 변경(application, domain, infrastructure, ui)
