+++ src/main/java/roomescape/reservation/application/dto/AvailableReservationTime.java
@@ -1,0 +1,11 @@
+package roomescape.reservation.application.dto;
+
+import java.time.LocalTime;
+
+public record AvailableReservationTime(
+        Long id,
+        LocalTime startAt,
+        boolean alreadyBooked
+) {
+
+}
