--- src/main/java/roomescape/reservation/controller/dto/AvailableReservationTimeRequest.java
@@ -1,9 +1,0 @@
-package roomescape.reservation.controller.dto;
-
-import java.time.LocalDate;
-
-public record AvailableReservationTimeRequest(
-        LocalDate date,
-        Long themeId
-) {
-}
