--- src/main/java/roomescape/reservation/controller/dto/AvailableReservationTimeResponse.java
@@ -1,10 +1,0 @@
-package roomescape.reservation.controller.dto;
-
-import java.time.LocalTime;
-
-public record AvailableReservationTimeResponse(
-        Long timeId,
-        LocalTime startAt,
-        boolean alreadyBooked
-) {
-}
