--- src/main/java/roomescape/reservation/controller/dto/CreateReservationRequest.java
@@ -1,12 +1,0 @@
-package roomescape.reservation.controller.dto;
-
-import java.time.LocalDate;
-
-public record CreateReservationRequest(
-        String name,
-        LocalDate date,
-        Long timeId,
-        Long themeId
-) {
-
-}
