--- src/main/java/roomescape/reservation/controller/dto/CreateReservationResponse.java
@@ -1,23 +1,0 @@
-package roomescape.reservation.controller.dto;
-
-import java.time.LocalDate;
-import roomescape.reservation.domain.Reservation;
-import roomescape.theme.controller.dto.ThemeResponse;
-
-public record CreateReservationResponse(
-        Long id,
-        String name,
-        LocalDate date,
-        ReservationTimeResponse time,
-        ThemeResponse theme
-) {
-
-    public static CreateReservationResponse from(final Reservation reservation) {
-        return new CreateReservationResponse(
-                reservation.getId(),
-                reservation.getName(),
-                reservation.getDate(),
-                ReservationTimeResponse.from(reservation.getTime()),
-                ThemeResponse.from(reservation.getTheme()));
-    }
-}
