--- src/main/java/roomescape/reservation/controller/dto/ReservationTimeRequest.java
@@ -1,10 +1,0 @@
-package roomescape.reservation.controller.dto;
-
-import com.fasterxml.jackson.annotation.JsonFormat;
-import java.time.LocalTime;
-
-public record ReservationTimeRequest(
-        @JsonFormat(pattern = "HH:mm") LocalTime startAt
-) {
-
-}
