+++ src/main/java/roomescape/reservation/domain/ReservationTimeRepository.java
@@ -1,0 +1,18 @@
+package roomescape.reservation.domain;
+
+import java.time.LocalTime;
+import java.util.List;
+import java.util.Optional;
+
+public interface ReservationTimeRepository {
+
+    Long save(ReservationTime reservationTime);
+
+    Optional<ReservationTime> findById(Long id);
+
+    List<ReservationTime> findAllByStartAt(LocalTime startAt);
+
+    List<ReservationTime> findAll();
+
+    void deleteById(Long id);
+}
