--- src/main/java/roomescape/reservation/service/dto/AvailableReservationTime.java
@@ -1,10 +1,0 @@
-package roomescape.reservation.service.dto;
-
-import java.time.LocalTime;
-
-public record AvailableReservationTime(
-        Long id,
-        LocalTime startAt,
-        boolean alreadyBooked
-) {
-}
