+++ src/main/java/roomescape/reservation/ui/dto/AvailableReservationTimeRequest.java
@@ -1,0 +1,9 @@
+package roomescape.reservation.ui.dto;
+
+import java.time.LocalDate;
+
+public record AvailableReservationTimeRequest(
+        LocalDate date,
+        Long themeId
+) {
+}
