+++ src/main/java/roomescape/reservation/ui/dto/AvailableReservationTimeResponse.java
@@ -1,0 +1,10 @@
+package roomescape.reservation.ui.dto;
+
+import java.time.LocalTime;
+
+public record AvailableReservationTimeResponse(
+        Long timeId,
+        LocalTime startAt,
+        boolean alreadyBooked
+) {
+}
