+++ src/main/java/roomescape/reservation/ui/dto/CreateReservationRequest.java
@@ -1,0 +1,12 @@
+package roomescape.reservation.ui.dto;
+
+import java.time.LocalDate;
+
+public record CreateReservationRequest(
+        String name,
+        LocalDate date,
+        Long timeId,
+        Long themeId
+) {
+
+}
