+++ src/main/java/roomescape/reservation/ui/dto/CreateReservationResponse.java
@@ -1,0 +1,23 @@
+package roomescape.reservation.ui.dto;
+
+import java.time.LocalDate;
+import roomescape.reservation.domain.Reservation;
+import roomescape.theme.ui.dto.ThemeResponse;
+
+public record CreateReservationResponse(
+        Long id,
+        String name,
+        LocalDate date,
+        ReservationTimeResponse time,
+        ThemeResponse theme
+) {
+
+    public static CreateReservationResponse from(final Reservation reservation) {
+        return new CreateReservationResponse(
+                reservation.getId(),
+                reservation.getName(),
+                reservation.getDate(),
+                ReservationTimeResponse.from(reservation.getTime()),
+                ThemeResponse.from(reservation.getTheme()));
+    }
+}
