+++ src/main/java/roomescape/reservation/ui/dto/ReservationTimeRequest.java
@@ -1,0 +1,10 @@
+package roomescape.reservation.ui.dto;
+
+import com.fasterxml.jackson.annotation.JsonFormat;
+import java.time.LocalTime;
+
+public record ReservationTimeRequest(
+        @JsonFormat(pattern = "HH:mm") LocalTime startAt
+) {
+
+}
