+++ src/main/java/roomescape/reservation/ui/dto/ReservationTimeResponse.java
@@ -1,0 +1,18 @@
+package roomescape.reservation.ui.dto;
+
+import com.fasterxml.jackson.annotation.JsonFormat;
+import java.time.LocalTime;
+import roomescape.reservation.domain.ReservationTime;
+
+public record ReservationTimeResponse(
+        Long id,
+        @JsonFormat(pattern = "HH:mm") LocalTime startAt
+) {
+
+    public static ReservationTimeResponse from(ReservationTime reservationTime) {
+        return new ReservationTimeResponse(
+                reservationTime.getId(),
+                reservationTime.getStartAt()
+        );
+    }
+}
