--- src/main/java/roomescape/theme/controller/dto/ThemeRequest.java
@@ -1,8 +1,0 @@
-package roomescape.theme.controller.dto;
-
-public record ThemeRequest(
-        String name,
-        String description,
-        String thumbnail
-) {
-}
