--- src/main/java/roomescape/theme/controller/dto/ThemeResponse.java
@@ -1,20 +1,0 @@
-package roomescape.theme.controller.dto;
-
-import roomescape.theme.domain.Theme;
-
-public record ThemeResponse(
-        Long id,
-        String name,
-        String description,
-        String thumbnail
-) {
-
-    public static ThemeResponse from(final Theme theme) {
-        return new ThemeResponse(
-                theme.getId(),
-                theme.getName(),
-                theme.getDescription(),
-                theme.getThumbnail()
-        );
-    }
-}
