+++ src/main/java/roomescape/theme/ui/dto/ThemeRequest.java
@@ -1,0 +1,8 @@
+package roomescape.theme.ui.dto;
+
+public record ThemeRequest(
+        String name,
+        String description,
+        String thumbnail
+) {
+}
