+++ src/main/java/roomescape/theme/ui/dto/ThemeResponse.java
@@ -1,0 +1,20 @@
+package roomescape.theme.ui.dto;
+
+import roomescape.theme.domain.Theme;
+
+public record ThemeResponse(
+        Long id,
+        String name,
+        String description,
+        String thumbnail
+) {
+
+    public static ThemeResponse from(final Theme theme) {
+        return new ThemeResponse(
+                theme.getId(),
+                theme.getName(),
+                theme.getDescription(),
+                theme.getThumbnail()
+        );
+    }
+}
