--- src/main/java/roomescape/user/controller/UserController.java
@@ -1,18 +1,0 @@
-package roomescape.user.controller;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-
-@Controller
-public class UserController {
-
-    @GetMapping("/reservation")
-    public String reservation() {
-        return "reservation";
-    }
-
-    @GetMapping
-    public String home() {
-        return "index";
-    }
-}
