+++ src/test/java/roomescape/reservation/controller/ReservationRestControllerTest.java
@@ -18,7 +18,8 @@ import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
 import org.springframework.http.HttpStatus;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.test.annotation.DirtiesContext;
-import roomescape.reservation.controller.dto.AvailableReservationTimeResponse;
+import roomescape.reservation.ui.ReservationRestController;
+import roomescape.reservation.ui.dto.AvailableReservationTimeResponse;
 
 @SpringBootTest(webEnvironment = WebEnvironment.DEFINED_PORT)
 @DirtiesContext(classMode = DirtiesContext.ClassMode.AFTER_EACH_TEST_METHOD)
